Добавил выход и удаление аккаунта и также повторно просмотрел иисправил свой код

+++ backend/app/Http/Controllers/Api/UserController.php
@@ -28,7 +28,7 @@ class UserController extends Controller
             'password' => Hash::make($request->password),
         ]);
 
-        return response()->json(['status' => true, 'message' => 'Пользователь прошел базовую регистрацию', 'email' => $user->email]);
+        return response()->json(['message' => 'Пользователь прошел базовую регистрацию', 'email' => $user->email]);
     }
 
     public function verifyRegistration(VerifyRegistrationRequest $request)
@@ -38,17 +38,16 @@ class UserController extends Controller
         if(VerificationCode::verify($request->verify_code, $request->email)){
             $user->markEmailAsVerified();
 
-            return response()->json(['status' => true, 'message' => 'Пользователь подтвердил свой аккаунт', 'data' => $user]);
+            return response()->json(['message' => 'Пользователь подтвердил свой аккаунт']);
         }
 
-        return response()->json(['status' => false, 'message' => 'Введенный код не относится ни к одному из пользователей'], 404);
-
+        return response()->json(['message' => 'Введенный код не относится ни к одному из пользователей'], 422);
     }
 
     public function authorization(AuthorizationRequest $request)
     {
         if(!Auth::attempt($request->only(['email', 'password']))){
-            return response()->json(['status' => false, 'message' => 'Введенные данные не относятся к существующему аккаунту'], 401);
+            return response()->json(['message' => 'Введенные данные не относятся к существующему аккаунту'], 422);
         }
 
         $user = Auth::user();
@@ -60,7 +59,7 @@ class UserController extends Controller
             $token = $user->createToken('token', ['*'], now()->addDays(1))->plainTextToken;
         }
 
-        return response()->json(['status' => true, 'message' => 'Пользователь успешно авторизирован', 'user' => $user, 'token' => $token]);
+        return response()->json(['message' => 'Пользователь успешно авторизирован', 'token' => $token]);
     }
 
     public function passwordResetEmail(PasswordResetEmailRequest $request)
@@ -69,7 +68,7 @@ class UserController extends Controller
 
         VerificationCode::send($user->email);
 
-        return response()->json(['status' => true, 'message' => 'Код для восстановления пароля был отправлен на почту', 'email' => $user->email]);
+        return response()->json(['message' => 'Код для восстановления пароля был отправлен на почту', 'email' => $user->email]);
     }
 
     public function passwordResetVerify(PasswordResetVerifyRequest $request)
@@ -77,11 +76,11 @@ class UserController extends Controller
         $user = User::where('email', $request->email)->first();
 
         if(VerificationCode::verify($request->verify_code, $request->email, false)){
-            return response()->json(['status' => true, 'message' => 'Введенный код для восстановления пароля подтвержден',
+            return response()->json(['message' => 'Введенный код для восстановления пароля подтвержден',
                 'email' => $user->email, 'verify_code' => $request->verify_code]);
         }
 
-        return response()->json(['status' => false, 'message' => 'Введенный код не относится ни к одному из пользователей'], 404);
+        return response()->json(['message' => 'Введенный код не относится ни к одному из пользователей'], 422);
     }
 
     public function passwordReset(PasswordResetRequest $request)
@@ -93,23 +92,57 @@ class UserController extends Controller
                 'password' => Hash::make($request->password)
             ]);
 
-            return response()->json(['status' => true, 'message' => 'Пароль был успешно изменен', 'data' => $user]);
+            return response()->json(['message' => 'Пароль был успешно изменен']);
         }
 
-        return response()->json(['status' => false, 'message' => 'Использованный код не относится ни к одному из пользователей']);
+        return response()->json(['message' => 'Использованный код не относится ни к одному из пользователей'], 422);
     }
 
     public function verifyCodeResend(PasswordResetEmailRequest $request)
     {
+        VerifyModel::where([['verifiable', '=', $request->email],
+            ['expires_at', '<=', now()]])->delete();
+
         $codes = VerifyModel::where('verifiable', $request->email)->get();
 
-        foreach ($codes as $code){
+        if($codes->isEmpty())
+        {
+            return response()->json(['message' => 'Ранее запрошенных кодов нету'], 422);
+        }
+
+        foreach($codes as $code)
+        {
             $code->delete();
         }
 
         VerificationCode::send($request->email);
 
-        return response()->json(['status' => true, 'message' => 'Код для восстановления пароля был отправлен на почту',
-            'email' => $request->email]);
+        return response()->json(['message' => 'Повторный код был отправлен на почту',
+            'email' => $request->email]); // мб почту убрать надо
+    }
+
+    public function logout()
+    {
+        Auth::user()->tokens()->delete();
+
+        return response()->json(['message' => 'Пользователь вышел из аккаунта']);
+    }
+
+    public function deleteUser()
+    {
+        $user = Auth::user();
+
+        $user->inviteTokens()->delete();
+
+        foreach($user->spaces as $space)
+        {
+            $space->delete();
+        }
+
+        Auth::user()->tokens()->delete();
+
+        $user->delete();
+
+        return response()->json(['message' => 'Аккаунт удален']);
     }
 }
